% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock.R
\name{clock}
\alias{clock}
\title{emoji version of time}
\usage{
clock(time)
}
\arguments{
\item{time}{a `POSIXct` object}
}
\value{
an emoji clock that is the closest to the given time
}
\description{
emoji version of time
}
\examples{
\dontrun{
clock( Sys.time() )
}

}
