% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon.R
\name{day_in_synodic_cycle}
\alias{day_in_synodic_cycle}
\title{Calculate the number of days in the synodic month}
\usage{
day_in_synodic_cycle(date)
}
\arguments{
\item{date}{a date}
}
\description{
This uses the approximation described in the
[lunar phase wikipedia page](https://en.wikipedia.org/wiki/Lunar_phase), i.e.
the number of days since `1900/01/01` modulo the length of
a synodic month (`29.530588853` days)
}
\examples{
\dontrun{
day_in_synodic_cycle( today() )
}
}
