% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.R
\name{square}
\alias{square}
\alias{diamond}
\title{geometric emoji}
\usage{
square(size = c("small", "medium", "medium-small", "large"),
  color = c("white", "black"))

diamond(size = c("large", "small"), color = c("orange", "blue"))
}
\arguments{
\item{size}{size}

\item{color}{color}
}
\description{
geometric emoji
}
\details{
For `square`:
- `size` should be one of `c( "small", "medium", "medium-small", "large" )`
- `color` should be "white" or "black"

For `diamond`
- `size` should be "large" or "small"
- `color` should be "orange" or "blue"
}
\examples{
\dontrun{
square( "small", "black" )
square( "large", "white" )

diamond( "small", "orange")
}

}
