% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji.R
\name{ji}
\alias{ji}
\title{Find a single emoji}
\usage{
ji(keyword)
}
\arguments{
\item{keyword}{Either name or keyword. If more than one emoji has the
specified keyword, will pick one at random.}
}
\description{
Find a single emoji
}
\examples{
emo::ji("banana")
emo::ji("monkey")
}
