% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_completion.R
\name{ji_completion}
\alias{ji_completion}
\title{emoji completion}
\usage{
ji_completion(token)
}
\arguments{
\item{token}{start of an emoji alias}
}
\description{
emoji completion
}
\examples{
ji_completion( "key" )

}
