% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_rx.R
\name{ji_count}
\alias{ji_count}
\title{Count the number of emojis in a string}
\usage{
ji_count(string)
}
\arguments{
\item{string}{Input vector}
}
\value{
An integer vector
}
\description{
Vectorised over `string`
}
