% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_rx.R
\name{ji_detect}
\alias{ji_detect}
\title{Detect the presence or absence of emojis in a string}
\usage{
ji_detect(string)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible to one}
}
\value{
A logical vector
}
\description{
Vectorised over `string`
}
\seealso{
[stringr::str_detect()]
}
