% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_rx.R
\name{ji_extract}
\alias{ji_extract}
\alias{ji_extract_all}
\title{Extract emojis from a string}
\usage{
ji_extract(string)

ji_extract_all(string, simplify = FALSE)
}
\arguments{
\item{string}{Input vector.}

\item{simplify}{see [stringr::str_extract_all()]}
}
\value{
A character vector
}
\description{
vectorised over `string`
}
\seealso{
[stringr::str_extract()] and [stringr::str_extract_all()]
}
