% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_fisher.R
\name{ji_fisher}
\alias{ji_fisher}
\title{Convert text to emoji (by letter)}
\usage{
ji_fisher(x)
}
\arguments{
\item{x}{Character string. Text you'd like to be emojified.}
}
\value{
Character string of emoji.
}
\description{
Convert text to emoji (by letter)
}
\examples{
ji_fisher("Carrie Fisher is a delight")
}
