% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_glue.R
\name{ji_glue}
\alias{ji_glue}
\title{emoji replacement}
\usage{
ji_glue(txt)
}
\arguments{
\item{txt}{character vector in which to replace `:something:` by an
emoji for `something`}
}
\description{
emoji replacement
}
\details{
[jitsu] is used on each pattern wrapped in `:`
}
\examples{
\dontrun{
  ji_glue( "My :cat,face,love: loves to :sleep,zzz:" )
}
}
