% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_rx.R
\name{ji_locate}
\alias{ji_locate}
\alias{ji_locate_all}
\title{Lodate the positio of emojis in a string}
\usage{
ji_locate(string)

ji_locate_all(string)
}
\arguments{
\item{string}{Input vector}
}
\value{
For `ji_locate` an integer matrix, for `ji_locate_all` a list
  of integer matrices
}
\description{
Vectorised over `string`
}
