% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_rx.R
\name{ji_match}
\alias{ji_match}
\alias{ji_match_all}
\title{Extract emojis from a string}
\usage{
ji_match(string)

ji_match_all(string)
}
\arguments{
\item{string}{Input vector}
}
\value{
see [stringr::str_match()]
}
\description{
Vectorized over `string`
}
\seealso{
[stringr::str_match]
}
