% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_p.R
\name{ji_p}
\alias{ji_p}
\title{Summarise your p-values with emoji}
\usage{
ji_p(x)
}
\arguments{
\item{x}{A vector of p-values.}
}
\description{
Summarise your p-values with emoji
}
\examples{
emo::ji_p(1)
emo::ji_p(0.1)
emo::ji_p(0.05)
emo::ji_p(0.01)
emo::ji_p(1e-6)

emo::ji_p(rbeta(50, 2, 5))
}
