% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_rx.R
\name{ji_replace}
\alias{ji_replace}
\alias{ji_replace_all}
\title{Replace emojis in a string}
\usage{
ji_replace(string, replacement)

ji_replace_all(string, replacement)
}
\arguments{
\item{string}{Input vector}

\item{replacement}{A character vector of replacements. Should either be
of length 1 or the same length as `string`. See [stringr::str_replace()] for details}
}
\value{
A character vector
}
\description{
Vectorised over `string` and `replacement`
}
