% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_rx.R
\name{ji_subset}
\alias{ji_subset}
\alias{ji_which}
\title{Keep strings containing an emoji, or find positions}
\usage{
ji_subset(string)

ji_which(string)
}
\arguments{
\item{string}{input vector}
}
\value{
A character vector
}
\description{
Keep strings containing an emoji, or find positions
}
\seealso{
[stringr::str_subset()]
}
