% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jis.R
\docType{data}
\name{jis}
\alias{jis}
\title{full list of emojis}
\format{tibble with columns
- id: identifier
- emoji: character representation of the emoji
- name: name
- category: category, e.g. "Smileys & People"
- subcategory: sub category, e.g. "face-positive"
- keywords: vector of keywords
- skin_tone: when applicable the skin tone, otherwise NA
- runes: vector of unicode runes, i.e. hexadecimal representations prefixed with "U+"
- apple ... windows: logical indicating if the given vendor supports the emoji}
\source{
[Unicode® Emoji Charts v5.0](http://unicode.org/emoji/charts/index.html)
}
\usage{
jis
}
\description{
full list of emojis
}
\keyword{datasets}
