% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keycap.R
\name{keycap}
\alias{keycap}
\title{Keycap emoji sequence}
\usage{
keycap(x)
}
\arguments{
\item{x}{character to emoji keycap}
}
\value{
a keycap version of `x`
}
\description{
Keycap emoji sequence
}
\examples{
\dontrun{
  keycap(3)
  keycap(10)
  keycap('#')
}

}
