% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medal.R
\name{medal}
\alias{medal}
\title{medals}
\usage{
medal(position)
}
\arguments{
\item{position}{1, 2, 3, or `first`, `second`, `third`, or `gold`, `silver`, `bronze`}
}
\description{
medals
}
\examples{
\dontrun{
medal(gold)
medal(third)
medal(2)
}

}
