% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon.R
\name{moon}
\alias{moon}
\title{moon phase}
\usage{
moon(date, day = day_in_synodic_cycle(date))
}
\arguments{
\item{date}{a date}

\item{day}{number of days since new moon

If not supplied, `day` is calculated using the
approximation of [day_in_synodic_cycle], i.e the number of
days since a known new moon modulo `29.530588853` days}
}
\value{
a moon emoji
}
\description{
moon phase
}
\examples{
\dontrun{
moon( today() )
}
}
