% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGLU.R
\name{runEMO}
\alias{runEMO}
\title{Main function to run GLU from the command line}
\usage{
runEMO(opt)
}
\arguments{
\item{opt}{List containing command line arguments.}
}
\description{
Main function to run GLU from the command line
}
