% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEMOForDirectory.R
\name{runEMOForDirectory}
\alias{runEMOForDirectory}
\title{Run EMO for all files in a directory}
\usage{
runEMOForDirectory(
  indir,
  outdir = NULL,
  imputeX = FALSE,
  lowthreshold = NULL,
  highthreshold = NULL,
  save = FALSE,
  saveevents = FALSE
)
}
\arguments{
\item{indir}{Path of input directory containing input files.}

\item{outdir}{Path of output directory where derived data will be stored.}

\item{imputeX}{Logical.}

\item{lowthreshold}{Numeric.}

\item{highthreshold}{Numeric.}

\item{save}{Logical. If TRUE then derived CGM sequence(s) are stored.}
}
\description{
Run EMO for all files in a directory
}
