% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEMOForFiles.R
\name{runEMOForFiles}
\alias{runEMOForFiles}
\title{Run EMO for given set of files}
\usage{
runEMOForFiles(
  files,
  indir,
  outdir = NULL,
  imputeX = FALSE,
  lowthreshold = NULL,
  highthreshold = NULL,
  save = FALSE,
  saveevents = FALSE
)
}
\arguments{
\item{files}{emotion data files for which variables will be derived.}

\item{indir}{Path of input directory containing input files.}

\item{outdir}{Path of output directory where derived data will be stored.}

\item{imputeX}{Logical. If TRUE then the XXXX approach to dealing with missing data is used.}

\item{lowthreshold}{Numeric.}

\item{highthreshold}{Numeric.}

\item{save}{Logical. If TRUE then derived emotion sequence(s) are stored.}
}
\description{
Run EMO for given set of files
}
