% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.R
\name{square}
\alias{square}
\title{square emoji}
\usage{
square(size = c("small", "medium", "medium-small", "large"),
  color = c("white", "black"))
}
\arguments{
\item{size}{size of the square ("small", "medium", "medium-small" or "large")}

\item{color}{"white" or "black"}
}
\description{
square emoji
}
\examples{
\dontrun{
square( "small", "black" )
square( "large", "white" )
}

}
