% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji.R
\name{ji_find}
\alias{ji_find}
\title{List all emoji with a given keyword}
\usage{
ji_find(keyword)
}
\arguments{
\item{keyword}{Emoji keyword}
}
\description{
Note that this is unlikely to print correctly on your R console, but
it will work in (e.g.) the RStudio viewer.
}
\examples{
emo::ji_find("happy")
}
