% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_glue.R
\name{ji_glue}
\alias{ji_glue}
\title{emoji glue}
\usage{
ji_glue(txt)
}
\arguments{
\item{txt}{character vector in which to replace `:something:` by an
emoji for `something`. see [ji()] for what `something` can be.

This is insipred from [glue::glue()] but uses `:` instead of braces
because that's typically how emojis are
[expressed](https://www.webpagefx.com/tools/emoji-cheat-sheet/)}
}
\description{
emoji glue
}
\examples{
\dontrun{
  ji_glue( ":cat:s love to :zzz:" )
}
}
