% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ji_glue.R
\name{ji_glue}
\alias{ji_glue}
\title{emoji glue}
\usage{
ji_glue(..., .envir = parent.frame())
}
\arguments{
\item{\dots}{strings to format, where `:x:` is replaced by an emoji for "x", using [ji()] and `:y*:` is replaced by all
emojis that match "y", using [ji_find()].}

\item{.envir}{see [glue::glue()]}
}
\description{
emoji glue
}
\examples{
\dontrun{
  ji_glue("one :heart:")
  ji_glue("many :heart*:")
}
}
\seealso{
[glue::glue()] for how the strings are concatenated
}
