% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitsu.R
\name{ji_filter}
\alias{ji_filter}
\alias{jitsu}
\alias{ji_set}
\title{extract an emoji}
\usage{
ji_filter(...)

jitsu(...)

ji_set(...)
}
\arguments{
\item{`...`}{set of filters, see details}
}
\description{
extract an emoji
}
\details{
`...` can contain
- bare symbols or strings, to be looked for in the name or the keywords. Understood as regular expressions.
- calls to the pseudo-functions `name`, `category`, `subcategory`, `skin_tone`, `keyword`, `runes`. These accept regular expressions
  to match the relevant column, e.g. `name("^cat")` or `name( fixed("cat") )`
- calls to the pseudo-function `vendor` that accept one symbol of
  `apple`, `google`, `twitter`, `one`, `facebook`, `messenger`, `samsung`, `windows`
- Any other call are just used verbatim in `filter`

`ji_filter` reworks the `...` as above and eventually returns the results of [dplyr::filter] on [jis]

`jitsu` selects one match at random between the results of `ji_filter`
}
\examples{
\dontrun{

# look in name and keywords with simple syntax
jitsu(cat,face)
jitsu(tears)
jitsu(poop)
jitsu(sad,cat)
jitsu(monkey)

# look only in specific columns
jitsu(name("cat"))
jitsu(skin_tone("^light"))
jitsu(subcategory("person-role"), skin_tone("medium-dark"))
jitsu(subcategory("face-fantasy") )

ji_filter( monkey )
ji_filter( cat, face )

}

}
